+++ wp-slimstat.php
@@ -3,7 +3,7 @@
  * Plugin Name: SlimStat Analytics
  * Plugin URI: https://wp-slimstat.com/
  * Description: The leading web analytics plugin for WordPress
- * Version: 5.4.12
+ * Version: 5.4.13
  * Author: Jason Crouse, VeronaLabs
  * Text Domain: wp-slimstat
  * Domain Path: /languages
@@ -20,7 +20,7 @@ if (!file_exists(__DIR__ . '/vendor/autoload.php')) {
 }
 
 // Set the plugin version and directory
-define('SLIMSTAT_ANALYTICS_VERSION', '5.4.12');
+define('SLIMSTAT_ANALYTICS_VERSION', '5.4.13');
 define('SLIMSTAT_FILE', __FILE__);
 define('SLIMSTAT_DIR', __DIR__);
 define('SLIMSTAT_URL', plugins_url('', __FILE__));
@@ -1230,10 +1230,11 @@ class wp_slimstat
 
             // Notices
             // -----------------------------------------------------------------------
-            'notice_latest_news' => 'on',
-            'notice_browscap'    => 'on',
-            'notice_geolite'     => 'on',
-            'notice_caching'     => 'on',
+            'notice_latest_news'       => 'on',
+            'notice_browscap'          => 'on',
+            'notice_browscap_fileinfo' => 'on',
+            'notice_geolite'           => 'on',
+            'notice_caching'           => 'on',
 
             // Network-wide Settings
             'locked_options' => '',